backend/requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
bitarray==3.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.2
cytoolz==1.0.1
dataclasses-json==0.6.7
dnspython==2.7.0
etelemetry==0.3.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
eval_type_backport==0.2.2
fastapi==0.115.11
filelock==3.17.0
fitz==0.0.1.dev2
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
google==3.0.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jwt==1.3.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-huggingface==0.1.2
langchain-mistralai==0.2.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mistralai==1.5.0
motor==3.7.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
passlib==1.7.4
pathlib==1.0.1
pillow==11.1.0
propcache==0.3.0
prov==2.0.1
puremagic==1.28
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
pymongo==4.11.2
PyMuPDF==1.25.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyunormalize==16.0.0
pywin32==308
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.49.0
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0

